Add tests_rest_server.cpp
We chose to start a server to test it directly.
We added a ServeThreaded method to rest_server.cpp.
We added a header-only httplib.h (https://github.com/yhirose/cpp-httplib) to create a simple client to test the server.

TODO: split tests by route. But starting multiple rest servers (and in particular Http::Endpoint), one after the other and with shuting down gives "Adress already in use".